+++ src/aosm/azext_aosm/deploy/artifact_manifest.py
@@ -1,7 +1,6 @@
 # Copyright (c) Microsoft Corporation. All rights reserved.
 # Highly Confidential Material
 """A module to handle interacting with artifact manifests."""
-
 from functools import cached_property, lru_cache
 from typing import Any, List, Union
 
@@ -64,7 +63,6 @@ class ArtifactManifestOperator:
             username=self._manifest_credentials["username"],
             password=self._manifest_credentials["acr_token"],
         )
-
         return client
 
     def _get_artifact_list(self) -> List[Artifact]:
@@ -100,6 +98,7 @@ class ArtifactManifestOperator:
                         artifact_type=artifact.artifact_type,
                         artifact_version=artifact.artifact_version,
                         artifact_client=self._get_artifact_client(artifact),
+                        manifest_credentials=self._manifest_credentials,
                     )
                 )